fix: vendor tree-kill
The `tree-kill` package has an RCE issue because the `pid` input is
passed unsanitised to an `exec` command. To work around this, we are
vendoring the code and adding some basic checks to make sure the pid is
valid.

Also adding a `.npmrc` file to prevent `package-lock.json` being
generated, and moved `debug` to prod deps because its used in
`lib/`