fix: merge scopes (configurations)
https://snyksec.atlassian.net/browse/TARDIS-1082

Discovered a bug when merging scopes together in the Scala scripts.

We have been doing `map ++ map`.
Which naively combines to maps ultimately setting the root element to the last value.
So that if a 'provided' scope is listed last, the root element only lists provided dependencies.
Implementing a more sophisticated function that combines both the Map and the Set values.

Co-authored-by: Johnny Dunsby <Jdunsby@users.noreply.github.com>
Co-authored-by: ola magdziarek <magdziarek@users.noreply.github.com>