fix: handle edge cases in stdout buffering
Fixes the tests and edge cases for how node buffers terminal output and large lines. (This
parsing difference was always present for debugging, generating spurious newlines.)

This version should not need to do any extra work rescanning the large json lines of the snyk scala plugin
for newlines nor print anything additional or in duplicate to the debug logging since it buffers each last
line of output until it is complete.